:recycle: refactor: Cell, Offset OffsetBase 상속받도록
- flutter 따라했음
- Cell이 Offset을 구성해도 되지만 cell.x, cell.y 처럼 사용하고 싶어서
