+++ src/main/java/tetro/Offset.java
@@ -1,30 +1,12 @@
 package tetro;
 
-import java.util.Objects;
-
 /**
  * 좌표의 x, y 위치를 나타낸다
  */
-public final class Offset {
-    public final int x;
-    public final int y;
+public final class Offset extends OffsetBase {
 
     public Offset(int x, int y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Offset)) return false;
-        Offset that = (Offset) o;
-        return this.x == that.x && this.y == that.y;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(x, y);
+        super(x, y);
     }
 
     @Override
